finished up code

+++ kahlifar-bot/helpers/send.js
@@ -3,17 +3,21 @@ const { sleep } = require(`${process.cwd()}/helpers/sleep.js`)
 const data = require(`${process.cwd()}/properties.json`)
 
 exports.sendError = async (message, content, doDelete, ephemeral) => {
-    var msg =  {}
+    var msg = {}
     let errorEmbed = new MessageEmbed()
         .setColor(data.helpers.send.colors.error)
         .setTitle(data.helpers.send.prefixTitle.error)
         .setDescription(content)
     if (message.type == "APPLICATION_COMMAND" || message.type == "MESSAGE_COMPONENT") {
-        msg = await message.reply({ embeds: [errorEmbed], fetchReply: true })
+        if (message.deferred) {
+            msg = message.editReply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        } else {
+            msg = await message.reply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        }
     } else {
-        msg = await message.channel.send({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        msg = await message.channel.send({ embeds: [errorEmbed], fetchReply: true })
     }
-    if (doDelete) {
+    if (doDelete && !ephemeral) {
         await sleep(data.deleteTime)
         msg.delete()
     }
@@ -26,12 +30,16 @@ exports.sendInfo = async (message, content, doDelete, ephemeral) => {
         .setColor(data.helpers.send.colors.info)
         .setTitle(data.helpers.send.prefixTitle.info)
         .setDescription(content)
-    if (message.type == "APPLICATION_COMMAND") {
-        msg = await message.reply({ embeds: [infoEmbed], fetchReply: true, ephemeral: ephemeral })
+    if (message.type == "APPLICATION_COMMAND" || message.type == "MESSAGE_COMPONENT") {
+        if (message.deferred) {
+            msg = message.editReply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        } else {
+            msg = await message.reply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        }
     } else {
         msg = await message.channel.send({ embeds: [infoEmbed], fetchReply: true })
     }
-    if (doDelete) {
+    if (doDelete && !ephemeral) {
         await sleep(data.deleteTime)
         msg.delete()
     }
@@ -39,17 +47,21 @@ exports.sendInfo = async (message, content, doDelete, ephemeral) => {
 
 
 exports.sendWarn = async (message, content, doDelete, ephemeral) => {
-    var msg =  {}
+    var msg = {}
     let errorEmbed = new MessageEmbed()
         .setColor(data.helpers.send.colors.warning)
         .setTitle(data.helpers.send.prefixTitle.warning)
         .setDescription(content)
-    if (message.type == "APPLICATION_COMMAND") {
-        msg = await message.reply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+    if (message.type == "APPLICATION_COMMAND" || message.type == "MESSAGE_COMPONENT") {
+        if (message.deferred) {
+            msg = message.editReply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        } else {
+            msg = await message.reply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        }
     } else {
         msg = await message.channel.send({ embeds: [errorEmbed], fetchReply: true })
     }
-    if (doDelete) {
+    if (doDelete && !ephemeral) {
         await sleep(data.deleteTime)
         msg.delete()
     }
@@ -57,17 +69,22 @@ exports.sendWarn = async (message, content, doDelete, ephemeral) => {
 
 
 exports.sendSuccess = async (message, content, doDelete, ephemeral = false) => {
-    var msg =  {}
+    var msg = {}
     let errorEmbed = new MessageEmbed()
         .setColor(data.helpers.send.colors.success)
         .setTitle(data.helpers.send.prefixTitle.success)
         .setDescription(content)
-    if (message.type == "APPLICATION_COMMAND") {
-        msg = await message.reply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+    if (message.type == "APPLICATION_COMMAND" || message.type == "MESSAGE_COMPONENT") {
+        if (message.deferred) {
+            msg = message.editReply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        } else {
+            msg = await message.reply({ embeds: [errorEmbed], fetchReply: true, ephemeral: ephemeral })
+        }
     } else {
         msg = await message.channel.send({ embeds: [errorEmbed], fetchReply: true })
     }
-    if (doDelete) {
+    if (doDelete && !ephemeral) {
+        if (!msg.deletable) return;
         await sleep(data.deleteTime)
         msg.delete()
     }
